Properly link with distribution-specific version of boost libraries
libprl_sdk links with multithreaded version of boost libraries, while
libprl_xml_model was using the default (non-mt) version of the same
libraries. This resulted in runtime linking of both (mt and non-mt)
versionis of boost libraries to libprl_sdk:

$ ldd /usr/lib64/libprl_sdk.so | grep boost
    libboost_filesystem-mt.so.1.53.0 => /lib64/libboost_filesystem-mt.so.1.53.0 (0x00007fea7bb9e000)
    libboost_system-mt.so.1.53.0 => /lib64/libboost_system-mt.so.1.53.0 (0x00007fea7b99a000)
    libboost_filesystem.so.1.53.0 => /lib64/libboost_filesystem.so.1.53.0 (0x00007fea79f0a000)
    libboost_system.so.1.53.0 => /lib64/libboost_system.so.1.53.0 (0x00007fea79d05000)

Choose the boost libraries version for dynamic linking the same way
as it is done in libprl_sdk.

Signed-off-by: Igor Bazhitov <ibazhitov@virtuozzo.com>