Refactor: move CVmEvent*, BaseNode to libprlcommon
The problem is that CVmEvent is a generic class, which is used everywhere in
prl projects. And if one wants to write a common function, which manipulates
CVmEvent - he adds a dependency of libprlcommon on libprlxmlmodel. Actually, it
was already done for ErrorSimple's method convertToEvent().

At the same time - there is no reason to keep this class in xmlmodel - because
its autogeneration is highly customized and the class itself is pretty stable.
So, let's move it to libprlcommon.