MaxPort should not be mandatory paramter #PSBM-66556
Currently MaxPort does not have "minOccurs", thus it makes it a mandatory
parameter in the config file. After the Update4 installation dispatcher.xml
was regenerated because MaxPort was obviously missing. Making minOccurs=0
makes it non-mandatory, and dispatcher will add the value to config on
startup automatically without regenerating the file entirely.