chipset alias
we have so called machine types for qemu vms which is a encoded
description of features that the virtual hardware provides. it is
an obscure string thus it's of no use to interpretate or parse it
in any manner.
let's have a place inside dispatcher config for a vm to hold that
string.