bug #PSBM-89791 aka Failed to read /vz/private/39b40805-7df1-4cae-adae-e318635f66fb/root. Exit code: 162
we need to convert paths inside a ct config from the ct backup to
relative ones in respect to the ct private folder path. due to
some unknown reason the code refuses to work if the private path
doesn't exist though it is designed to work even at a host different
from the one where a potential folder resides.
drop the unsuitable test.