Add equality condition for gateway ipv6 PSBM-139469.
The condition for the hotplug update is to check that the adapter parameters have been changed.
The operator== in the CVmGenericNetworkAdapter class is responsible for such a check.
However, operator== for this class did not include a condition for gateway with ipv6,
and when changing ipv6 gateway, the configuration was not updated.