Schema/VmConfig.xsd: Do not change nvram path in dispatcher if vm was created by virt-install #PSBM-141223
When creating a VM with an nvram disk using the virt-install utility, libvirt
puts the nvram of the created vm in a special directory /var/lib/qemu/nvram/.
If there is a dispatcher in the system, it in turn changes config.pvs and saves paths to
nvram disk relative to VmHome, after changing config.pvs the dispatcher sends an event
in libvirt, in the end we get that in the libvirt config the paths to nvram are built
regarding the VmHome dispatcher. This fix prevents changes to the nvram path to the disk if vm with
nvram was created by libvirt.

Signed-off-by: Efim Shevrin <efim.shevrin@virtuozzo.com>